Bug 1874684 - Part 18: Fix narrowing warnings from clang-tidy. r=allstarschh
Differential Revision: https://phabricator.services.mozilla.com/D198552

UltraBlame original commit: 61e7759827c23a2aa53b1977347c9d9a65158ef0

+++ js/src/builtin/temporal/Int96.cpp
@@ -39,12 +39,12 @@ mozilla::Maybe<Int96> Int96::fromInteger(double value) {
 
   
   
-  constexpr size_t DigitBits = 32;
+  constexpr int DigitBits = 32;
 
   
   Int96::Digits digits = {};
 
-  int exponent = mozilla::ExponentComponent(value);
+  int exponent = int(mozilla::ExponentComponent(value));
   MOZ_ASSERT(0 <= exponent && exponent <= 95,
              "exponent is lower than exponent of 0x1p+96");
 
@@ -62,7 +62,7 @@ mozilla::Maybe<Int96> Int96::fromInteger(double value) {
   int msdTopBit = exponent % DigitBits;
 
   
-  int remainingMantissaBits = Double::kSignificandWidth - msdTopBit;
+  int remainingMantissaBits = int(Double::kSignificandWidth - msdTopBit);
   digits[--length] = mantissa >> remainingMantissaBits;
